chore(rich-text-input): move to packages
