+++ packages/components/inputs/rich-text-input/README.md
