+++ packages/components/inputs/rich-text-input/src/editor.js
@@ -9,9 +9,9 @@ import Stack from '@commercetools-uikit/spacings-stack';
 import { AngleUpIcon, AngleDownIcon } from '@commercetools-uikit/icons';
 import Constraints from '@commercetools-uikit/constraints';
 import FlatButton from '@commercetools-uikit/flat-button';
-import RichTextBody from '../../internals/rich-text-body';
-import HiddenInput from '../../internals/rich-text-body/hidden-input';
-import messages from '../../internals/messages/multiline-input';
+import RichTextBody from '../../../../../src/components/internals/rich-text-body';
+import HiddenInput from '../../../../../src/components/internals/rich-text-body/hidden-input';
+import messages from '../../../../../src/components/internals/messages/multiline-input';
 import { EditorWrapper } from './editor.styles';
 
 const COLLAPSED_HEIGHT = 32;
