+++ packages/components/inputs/rich-text-input/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './rich-text-input';
+
+export { default as version } from './version';
