+++ packages/components/inputs/rich-text-input/src/rich-text-input.spec.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { screen, render } from '../../../../test/test-utils';
+import { screen, render } from '../../../../../test/test-utils';
 import RichTextInput from './rich-text-input';
 
 // mocks
