+++ packages/components/inputs/rich-text-input/src/rich-text-input.visualroute.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Switch, Route } from 'react-router';
 import { RichTextInput } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 const minimalValue = 'Hello World';
