--- src/components/inputs/rich-text-input/index.js
@@ -1,1 +1,0 @@
-export { default } from './rich-text-input';
